fair_queue: Move io_throttler to io_queue.hh
Now when all outgoing capacity throttling code is sitting in the
IO-queue, it's time to move the io_throttler class itself.
