+++ tests/perf/fair_queue_perf.cc
@@ -84,7 +84,7 @@ future<> perf_fair_queue::test(bool loc) {
 
     auto invokers = local_fq.invoke_on_all([loc] (local_fq_and_class& local) {
         return parallel_for_each(std::views::iota(0u, requests_to_dispatch), [&local, loc] (unsigned dummy) {
-            auto cap = io_throttler::capacity_t(1);
+            auto cap = fair_queue_entry::capacity_t(1);
             auto req = std::make_unique<local_fq_entry>(cap, [&local, loc, cap] {
                 local.executed++;
                 local.queue(loc).notify_request_finished(cap);
